aivants comments
